+++ src/wallpaper/constants.clj
@@ -2,6 +2,6 @@
   "This namespace is for any constants we might have for the application."
   (:gen-class))
 
-(def VERSION "0.1.5")
+(def VERSION "0.1.6")
 
 (def APP_NAME "wallpaper")